+++ __tests__/flat.js
@@ -1,11 +1,8 @@
-import chai from 'chai';
-import redraft, { createStylesRenderer } from '../src';
-import * as raws from './raws';
-import { joinRecursively, makeList } from './helpers';
-import ReactDOMServer from 'react-dom/server';
 import React from 'react';
-
-const should = chai.should();
+import ReactDOMServer from 'react-dom/server';
+import redraft, { createStylesRenderer } from '../src';
+import * as raws from './utils/raws';
+import { joinRecursively, makeList } from './utils/helpers';
 
 const customStyleMap = {
   BOLD: {
@@ -38,7 +35,7 @@ const customStyleMapReact = {
 };
 
 const stringifyStyles = reduced => Object.keys(reduced)
-  .map(key => `${key}:${reduced[key]};`).join('');
+  .map(key => `${key}:${reduced[key]}`).join(';');
 
 // render to HTML
 const styles = createStylesRenderer(
@@ -89,24 +86,29 @@ const renderersReact = {
   entities: entitiesReact,
 };
 
-// Helpers for te
-const bold = 'font-weight:bold;';
-const italic = 'font-style:italic;';
-const textDecoration = 'text-decoration:underline line-through;';
 
-const corretRender = `<p><span style="${bold}">Lorem </span><a href="http://zombo.com/"><span style="${bold}${italic}">ipsum</span></a><span style="${bold}${italic}${textDecoration}"> dolor</span><span style="${italic}"> sit amet,</span> pro nisl sonet ad. </p><blockquote>Eos affert numquam id, in est meis nobis. Legimus singulis suscipiantur eum in, <span style="${italic}">ceteros invenire </span>tractatos his id. </blockquote><p><span style="${bold}">Facer facilis definiebas ea pro, mei malis libris latine an. Senserit moderatius vituperata vis in.</span></p>` // eslint-disable-line max-len
+// Helpers for te
+const bold = 'font-weight:bold';
+const italic = 'font-style:italic';
+const textDecoration = 'text-decoration:underline line-through';
+
+const getStyles = (...arr) => arr.join(';');
+
+const correctRender = `<p><span style="${bold}">Lorem </span><a href="http://zombo.com/"><span style="${getStyles(bold, italic)}">ipsum</span></a><span style="${getStyles(bold, italic, textDecoration)}"> dolor</span><span style="${italic}"> sit amet,</span> pro nisl sonet ad. </p><blockquote>Eos affert numquam id, in est meis nobis. Legimus singulis suscipiantur eum in, <span style="${italic}">ceteros invenire </span>tractatos his id. </blockquote><p><span style="${bold}">Facer facilis definiebas ea pro, mei malis libris latine an. Senserit moderatius vituperata vis in.</span></p>` // eslint-disable-line max-len
 
 
 describe('redraft with flat styles', () => {
-  it('should render flat styles correctly', () => {
+  test('should render flat styles correctly', () => {
     const rendered = redraft(raws.raw, renderers);
     const joined = joinRecursively(rendered);
-    joined.should.equal(corretRender);
+    expect(joined).toBe(correctRender);
   });
-  it('should render flat styles correctly with ReactDOMServer.renderToStaticMarkup', () => {
-    const rendered = <div>{redraft(raws.raw, renderersReact)}</div>;
-    const stringified = ReactDOMServer.renderToStaticMarkup(rendered);
-    stringified.should.equal(`<div>${corretRender}</div>`
-    );
-  });
+  test(
+    'should render flat styles correctly with ReactDOMServer.renderToStaticMarkup',
+    () => {
+      const rendered = <div>{redraft(raws.raw, renderersReact)}</div>;
+      const stringified = ReactDOMServer.renderToStaticMarkup(rendered);
+      expect(stringified).toBe(`<div>${correctRender}</div>`);
+    }
+  );
 });