+++ __tests__/utils/raws.js
@@ -83,6 +83,18 @@ export const raw2 = {
   }],
 };
 
+export const unsportedType = {
+  entityMap: {},
+  blocks: [{
+    key: 'az45a',
+    text: 'Block of this type is not supported', // eslint-disable-line max-len
+    type: 'custom',
+    depth: 0,
+    inlineStyleRanges: [],
+    entityRanges: [],
+  }],
+};
+
 export const raw3 = {
   entityMap: {},
   blocks: [{
@@ -92,7 +104,6 @@ export const raw3 = {
     depth: 0,
     inlineStyleRanges: [],
     entityRanges: [],
-    data: {},
   }, {
     key: '520kr',
     text: 'A quote',
@@ -100,7 +111,6 @@ export const raw3 = {
     depth: 0,
     inlineStyleRanges: [],
     entityRanges: [],
-    data: {},
   }, {
     key: 'c3taj',
     text: 'Spanning multiple lines',
@@ -108,7 +118,6 @@ export const raw3 = {
     depth: 0,
     inlineStyleRanges: [],
     entityRanges: [],
-    data: {},
   }, {
     key: '6aaeh',
     text: 'A second paragraph.',
@@ -116,7 +125,6 @@ export const raw3 = {
     depth: 0,
     inlineStyleRanges: [],
     entityRanges: [],
-    data: {},
   }],
 };