+++ src/index.js
@@ -1,11 +1,15 @@
 import RawParser from './RawParser';
 import createStylesRenderer from './createStyleRenderer';
 import { renderNode, render } from './render';
+import CompositeDecorator from './helpers/CompositeDecorator';
+import createBlockRenderer from './createBlockRenderer';
 
 export {
   createStylesRenderer,
+  createBlockRenderer,
   RawParser,
   renderNode,
+  CompositeDecorator,
 };
 
 export default render;