+++ src/render.js
@@ -119,7 +119,7 @@ const byDepth = (blocks) => {
  * Conditionaly render a group if its not empty,
  * pass all the params to the renderers
  */
-const renderGroup = (group, blockRenderers, rendered, params) => {
+const renderGroup = (group, blockRenderers, rendered, params, options) => {
   const {
     prevType: type,
     prevDepth: depth,
@@ -130,12 +130,16 @@ const renderGroup = (group, blockRenderers, rendered, params) => {
   if (group.length === 0) {
     return;
   }
-  if (blockRenderers[type]) {
-    rendered.push(blockRenderers[type](group, {
+  const renderCb = blockRenderers[type] || blockRenderers[options.blockFallback];
+  if (renderCb) {
+    const props = {
       depth,
       keys,
-      data,
-    }));
+    };
+    if (data && data.some(item => !!item)) {
+      props.data = data;
+    }
+    rendered.push(renderCb(group, props));
     return;
   }
   rendered.push(group);
@@ -182,7 +186,8 @@ const renderBlocks = (blocks, inlineRenderers = {}, blockRenderers = {},
         group,
         blockRenderers,
         rendered,
-        { prevType, prevDepth, prevKeys, prevData }
+        { prevType, prevDepth, prevKeys, prevData },
+        options
       );
       // reset group vars
       // IDEA: might be worth to group those into an instance and just newup a new one
@@ -211,7 +216,8 @@ const renderBlocks = (blocks, inlineRenderers = {}, blockRenderers = {},
     group,
     blockRenderers,
     rendered,
-    { prevType, prevDepth, prevKeys, prevData }
+    { prevType, prevDepth, prevKeys, prevData },
+    options
   );
   return checkJoin(rendered, options);
 };
@@ -237,7 +243,7 @@ export const render = (raw, renderers = {}, options = {}) => {
     decorators,
   } = renderers;
   // If decorators are present, they are maped with the blocks array
-  const blocksWithDecorators = decorators || options.Decorator
+  const blocksWithDecorators = decorators
     ? withDecorators(raw, decorators, options)
     : raw.blocks;
   // Nest blocks by depth